WIP: Electronizing the app
GameMaster and TV are now windows of an Electron app.

The build/development process is U G L Y; right now, since we're using
ES6 on the server and Electron does not support ES6 by default, the
quickest way I got was to have a two-build step before actually starting
the app (build the web client [a bundle], build the server [using
babel] and then start Electron). This actually means that every change
requires this. The server is still able to be launched "standalone", but
the apps aren't since the bundle import of the code is now using `require`
since it resolves the paths correctly. I'm sure we'll be able to really
improve this, but to ship this we'll need to do this dirty.