Separating Webpack entrypoint from app
HACK: manually "copying" dependencies by declaring them in the
entrypoint. If I understand correctly the best practice would be to
configure Webpack properly so that tv.html and master.html can
themselves be the entrypoints. This commit is essentially to separate
the entrypoint from what's the main structure of the application.